Add back capa scale; review upgrade step
