+++ src/brasil/gov/portal/profiles/default/propertiestool.xml
@@ -41,6 +41,7 @@
   </object>
   <object name="imaging_properties" meta_type="Plone Property Sheet">
     <property name="allowed_sizes" type="lines" purge="False">
+      <element value="capa 230:230"/>
       <element value="galeria_de_album_thumb 238:133"/>
       <element value="galeria_de_foto_view 1150:650"/>
       <element value="Imagem-Full: 1150:1150"/>
